+++ ICD.Connect.Routing/Devices/RelaySwitcher/RelaySwitcherDevice.cs
@@ -1,0 +1,431 @@
+﻿using System;
+using System.Collections.Generic;
+using ICD.Common.Utils.EventArguments;
+using ICD.Common.Utils.Extensions;
+using ICD.Common.Utils.Services.Logging;
+using ICD.Connect.Devices.Controls;
+using ICD.Connect.Devices.EventArguments;
+using ICD.Connect.Protocol.Extensions;
+using ICD.Connect.Protocol.Ports.RelayPort;
+using ICD.Connect.Routing.Connections;
+using ICD.Connect.Routing.Controls;
+using ICD.Connect.Routing.EventArguments;
+using ICD.Connect.Routing.Utils;
+using ICD.Connect.Settings;
+
+namespace ICD.Connect.Routing.Devices.RelaySwitcher
+{
+	public sealed class RelaySwitcherDevice : AbstractRouteSwitcherDevice<RelaySwitcherDeviceSettings>
+	{
+		private const int INPUT_1_ADDRESS = 1;
+		private const int INPUT_2_ADDRESS = 2;
+		private const int OUTPUT_ADDRESS = 1;
+
+		private const bool INPUT_1_ACTIVE = true;
+		private const bool INPUT_2_ACTIVE = false;
+
+		private const eConnectionType CONNECTION_TYPE =
+			eConnectionType.Audio |
+			eConnectionType.Video |
+			eConnectionType.Usb;
+
+		/// <summary>
+		/// Raised when an input source status changes.
+		/// </summary>
+		public override event EventHandler<SourceDetectionStateChangeEventArgs> OnSourceDetectionStateChange;
+
+		/// <summary>
+		/// Raised when the device starts/stops actively using an input, e.g. unroutes an input.
+		/// </summary>
+		public override event EventHandler<ActiveInputStateChangeEventArgs> OnActiveInputsChanged;
+
+		/// <summary>
+		/// Called when a route changes.
+		/// </summary>
+		public override event EventHandler<RouteChangeEventArgs> OnRouteChange;
+
+		/// <summary>
+		/// Raised when the device starts/stops actively transmitting on an output.
+		/// </summary>
+		public override event EventHandler<TransmissionStateEventArgs> OnActiveTransmissionStateChanged;
+
+		private readonly SwitcherCache m_Cache;
+
+		private IRelayPort m_Port;
+
+		#region Properties
+
+		/// <summary>
+		/// Gets/sets the IO port for the switcher.
+		/// </summary>
+		public IRelayPort Port
+		{
+			get { return m_Port; }
+			set
+			{
+				if (value == m_Port)
+					return;
+
+				Unsubscribe(m_Port);
+				m_Port = value;
+				Subscribe(m_Port);
+
+				UpdateCachedOnlineStatus();
+				UpdateActiveInput();
+			}
+		}
+
+		#endregion
+
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		public RelaySwitcherDevice()
+		{
+			m_Cache = new SwitcherCache();
+			m_Cache.OnActiveInputsChanged += CacheOnActiveInputsChanged;
+			m_Cache.OnActiveTransmissionStateChanged += CacheOnActiveTransmissionStateChanged;
+			m_Cache.OnRouteChange += CacheOnRouteChange;
+			m_Cache.OnSourceDetectionStateChange += CacheOnSourceDetectionStateChange;
+		}
+
+		#region Methods
+
+		/// <summary>
+		/// Returns true if a signal is detected at the given input.
+		/// </summary>
+		/// <param name="input"></param>
+		/// <param name="type"></param>
+		/// <returns></returns>
+		public override bool GetSignalDetectedState(int input, eConnectionType type)
+		{
+			if (!ContainsInput(input))
+				throw new ArgumentOutOfRangeException("input");
+
+			if (type == eConnectionType.None)
+				throw new ArgumentOutOfRangeException("type");
+
+			return true;
+		}
+
+		/// <summary>
+		/// Gets the input at the given address.
+		/// </summary>
+		/// <param name="input"></param>
+		/// <returns></returns>
+		public override ConnectorInfo GetInput(int input)
+		{
+			if (!ContainsInput(input))
+				throw new ArgumentOutOfRangeException("input");
+
+			return new ConnectorInfo(input, CONNECTION_TYPE);
+		}
+
+		/// <summary>
+		/// Returns true if the destination contains an input at the given address.
+		/// </summary>
+		/// <param name="input"></param>
+		/// <returns></returns>
+		public override bool ContainsInput(int input)
+		{
+			return input == INPUT_1_ADDRESS || input == INPUT_2_ADDRESS;
+		}
+
+		/// <summary>
+		/// Returns the inputs.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<ConnectorInfo> GetInputs()
+		{
+			yield return GetInput(INPUT_1_ADDRESS);
+			yield return GetInput(INPUT_2_ADDRESS);
+		}
+
+		/// <summary>
+		/// Gets the output at the given address.
+		/// </summary>
+		/// <param name="output"></param>
+		/// <returns></returns>
+		public override ConnectorInfo GetOutput(int output)
+		{
+			if (!ContainsOutput(output))
+				throw new ArgumentOutOfRangeException("output");
+
+			return new ConnectorInfo(output, CONNECTION_TYPE);
+		}
+
+		/// <summary>
+		/// Returns true if the source contains an output at the given address.
+		/// </summary>
+		/// <param name="output"></param>
+		/// <returns></returns>
+		public override bool ContainsOutput(int output)
+		{
+			return output == OUTPUT_ADDRESS;
+		}
+
+		/// <summary>
+		/// Returns the outputs.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<ConnectorInfo> GetOutputs()
+		{
+			yield return GetOutput(OUTPUT_ADDRESS);
+		}
+
+		/// <summary>
+		/// Gets the input routed to the given output matching the given type.
+		/// </summary>
+		/// <param name="output"></param>
+		/// <param name="type"></param>
+		/// <returns></returns>
+		/// <exception cref="InvalidOperationException">Type has multiple flags.</exception>
+		public override ConnectorInfo? GetInput(int output, eConnectionType type)
+		{
+			if (!ContainsOutput(output))
+				throw new ArgumentOutOfRangeException("output");
+
+			if (type == eConnectionType.None)
+				throw new ArgumentOutOfRangeException("type");
+
+			return m_Cache.GetInputConnectorInfoForOutput(output, type);
+		}
+
+		/// <summary>
+		/// Gets the outputs for the given input.
+		/// </summary>
+		/// <param name="input"></param>
+		/// <param name="type"></param>
+		/// <returns></returns>
+		public override IEnumerable<ConnectorInfo> GetOutputs(int input, eConnectionType type)
+		{
+			if (!ContainsOutput(input))
+				throw new ArgumentOutOfRangeException("input");
+
+			if (type == eConnectionType.None)
+				throw new ArgumentOutOfRangeException("type");
+
+			return m_Cache.GetOutputsForInput(input, type);
+		}
+
+		/// <summary>
+		/// Performs the given route operation.
+		/// </summary>
+		/// <param name="info"></param>
+		/// <returns></returns>
+		public override bool Route(RouteOperation info)
+		{
+			if (info == null)
+				throw new ArgumentNullException("info");
+
+			int output = info.LocalOutput;
+			if (!ContainsOutput(output))
+				throw new ArgumentException("No output with address " + output);
+
+			int input = info.LocalInput;
+			return Route(input);
+		}
+
+		/// <summary>
+		/// Routes the given input to the output.
+		/// </summary>
+		/// <param name="input"></param>
+		/// <returns></returns>
+		public bool Route(int input)
+		{
+			bool open;
+
+			switch (input)
+			{
+				case INPUT_1_ADDRESS:
+					open = INPUT_1_ACTIVE;
+					break;
+
+				case INPUT_2_ADDRESS:
+					open = INPUT_2_ACTIVE;
+					break;
+
+				default:
+					throw new ArgumentOutOfRangeException("output");
+			}
+
+			m_Port.SetOpen(open);
+			return true;
+		}
+
+		/// <summary>
+		/// Stops routing to the given output.
+		/// </summary>
+		/// <param name="output"></param>
+		/// <param name="type"></param>
+		/// <returns>True if successfully cleared.</returns>
+		public override bool ClearOutput(int output, eConnectionType type)
+		{
+			// Clearing output doesn't really make sense
+			return false;
+		}
+
+		#endregion
+
+		#region Private Methods
+
+		/// <summary>
+		/// Gets the current online status of the device.
+		/// </summary>
+		/// <returns></returns>
+		protected override bool GetIsOnlineStatus()
+		{
+			return Port != null && Port.IsOnline;
+		}
+
+		/// <summary>
+		/// Updates the active input to match the IO port state.
+		/// </summary>
+		private void UpdateActiveInput()
+		{
+			bool open = Port != null && !Port.Closed;
+
+			m_Cache.SetInputForOutput(OUTPUT_ADDRESS, open == INPUT_1_ACTIVE ? INPUT_1_ADDRESS : INPUT_2_ADDRESS, CONNECTION_TYPE);
+		}
+
+		#endregion
+
+		#region Port Callbacks
+
+		/// <summary>
+		/// Subscribe to the port events.
+		/// </summary>
+		/// <param name="port"></param>
+		private void Subscribe(IRelayPort port)
+		{
+			if (port == null)
+				return;
+
+			port.OnIsOnlineStateChanged += PortOnIsOnlineStateChanged;
+			port.OnClosedStateChanged += PortOnClosedStateChanged;
+		}
+
+		/// <summary>
+		/// Unsubscribe from the port events.
+		/// </summary>
+		/// <param name="port"></param>
+		private void Unsubscribe(IRelayPort port)
+		{
+			if (port == null)
+				return;
+
+			port.OnIsOnlineStateChanged -= PortOnIsOnlineStateChanged;
+			port.OnClosedStateChanged -= PortOnClosedStateChanged;
+		}
+
+		/// <summary>
+		/// Called when the port online state changes.
+		/// </summary>
+		/// <param name="sender"></param>
+		/// <param name="eventArgs"></param>
+		private void PortOnIsOnlineStateChanged(object sender, DeviceBaseOnlineStateApiEventArgs eventArgs)
+		{
+			UpdateCachedOnlineStatus();
+		}
+
+		/// <summary>
+		/// Called when the port digital output state changes.
+		/// </summary>
+		/// <param name="sender"></param>
+		/// <param name="boolEventArgs"></param>
+		private void PortOnClosedStateChanged(object sender, BoolEventArgs boolEventArgs)
+		{
+			UpdateActiveInput();
+		}
+
+		#endregion
+
+		#region Cache Callbacks
+
+		private void CacheOnActiveInputsChanged(object sender, ActiveInputStateChangeEventArgs eventArgs)
+		{
+			OnActiveInputsChanged.Raise(this, new ActiveInputStateChangeEventArgs(eventArgs));
+		}
+
+		private void CacheOnActiveTransmissionStateChanged(object sender, TransmissionStateEventArgs eventArgs)
+		{
+			OnActiveTransmissionStateChanged.Raise(this, new TransmissionStateEventArgs(eventArgs));
+		}
+
+		private void CacheOnRouteChange(object sender, RouteChangeEventArgs eventArgs)
+		{
+			OnRouteChange.Raise(this, new RouteChangeEventArgs(eventArgs));
+		}
+
+		private void CacheOnSourceDetectionStateChange(object sender, SourceDetectionStateChangeEventArgs eventArgs)
+		{
+			OnSourceDetectionStateChange.Raise(this, new SourceDetectionStateChangeEventArgs(eventArgs));
+		}
+
+		#endregion
+
+		#region Settings
+
+		/// <summary>
+		/// Override to apply settings to the instance.
+		/// </summary>
+		/// <param name="settings"></param>
+		/// <param name="factory"></param>
+		protected override void ApplySettingsFinal(RelaySwitcherDeviceSettings settings, IDeviceFactory factory)
+		{
+			base.ApplySettingsFinal(settings, factory);
+
+			IRelayPort port = null;
+
+			if (settings.Port != null)
+			{
+				try
+				{
+					port = factory.GetPortById((int)settings.Port) as IRelayPort;
+				}
+				catch (KeyNotFoundException)
+				{
+					Logger.Log(eSeverity.Error, "No IO port with id {0}", settings.Port);
+				}
+			}
+
+			Port = port;
+		}
+
+		/// <summary>
+		/// Override to clear the instance settings.
+		/// </summary>
+		protected override void ClearSettingsFinal()
+		{
+			base.ClearSettingsFinal();
+
+			Port = null;
+		}
+
+		/// <summary>
+		/// Override to apply properties to the settings instance.
+		/// </summary>
+		/// <param name="settings"></param>
+		protected override void CopySettingsFinal(RelaySwitcherDeviceSettings settings)
+		{
+			base.CopySettingsFinal(settings);
+
+			settings.Port = Port == null ? (int?)null : Port.Id;
+		}
+
+		/// <summary>
+		/// Override to add controls to the device.
+		/// </summary>
+		/// <param name="settings"></param>
+		/// <param name="factory"></param>
+		/// <param name="addControl"></param>
+		protected override void AddControls(RelaySwitcherDeviceSettings settings, IDeviceFactory factory, Action<IDeviceControl> addControl)
+		{
+			base.AddControls(settings, factory, addControl);
+
+			addControl(new RouteSwitcherControl(this, 0));
+		}
+
+		#endregion
+	}
+}