+++ ICD.Connect.Routing/Devices/RelaySwitcher/RelaySwitcherDeviceSettings.cs
@@ -1,0 +1,42 @@
+﻿using ICD.Common.Utils.Xml;
+using ICD.Connect.Devices;
+using ICD.Connect.Protocol.Ports.RelayPort;
+using ICD.Connect.Settings.Attributes;
+using ICD.Connect.Settings.Attributes.SettingsProperties;
+
+namespace ICD.Connect.Routing.Devices.RelaySwitcher
+{
+	[KrangSettings("RelaySwitcher", typeof(RelaySwitcherDevice))]
+	public sealed class RelaySwitcherDeviceSettings : AbstractDeviceSettings
+	{
+		public const string PORT_ELEMENT = "Port";
+
+		/// <summary>
+		/// The port id.
+		/// </summary>
+		[OriginatorIdSettingsProperty(typeof(IRelayPort))]
+		public int? Port { get; set; }
+
+		/// <summary>
+		/// Writes property elements to xml.
+		/// </summary>
+		/// <param name="writer"></param>
+		protected override void WriteElements(IcdXmlTextWriter writer)
+		{
+			base.WriteElements(writer);
+
+			writer.WriteElementString(PORT_ELEMENT, Port == null ? null : IcdXmlConvert.ToString((int)Port));
+		}
+
+		/// <summary>
+		/// Updates the settings from xml.
+		/// </summary>
+		/// <param name="xml"></param>
+		public override void ParseXml(string xml)
+		{
+			base.ParseXml(xml);
+
+			Port = XmlUtils.TryReadChildElementContentAsInt(xml, PORT_ELEMENT);
+		}
+	}
+}